move materializer to its own thing
