+++ r/src/materialize.c
@@ -1,0 +1,39 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+#define R_NO_REMAP
+#include <R.h>
+#include <Rinternals.h>
+
+#include "nanoarrow.h"
+
+SEXP nanoarrow_materialize_chr(struct ArrowArrayView* array_view) {
+  SEXP result = PROTECT(Rf_allocVector(STRSXP, array_view->array->length));
+
+  struct ArrowStringView item;
+  for (R_xlen_t i = 0; i < array_view->array->length; i++) {
+    if (ArrowArrayViewIsNull(array_view, i)) {
+      SET_STRING_ELT(result, i, NA_STRING);
+    } else {
+      item = ArrowArrayViewGetStringUnsafe(array_view, i);
+      SET_STRING_ELT(result, i, Rf_mkCharLenCE(item.data, item.n_bytes, CE_UTF8));
+    }
+  }
+
+  UNPROTECT(1);
+  return result;
+}
